Convert nice_go into a new integration and restore linear_garage_door
