+++ CODEOWNERS
@@ -968,6 +968,8 @@ build.json @home-assistant/supervisor
 /tests/components/nfandroidtv/ @tkdrob
 /homeassistant/components/nibe_heatpump/ @elupus
 /tests/components/nibe_heatpump/ @elupus
+/homeassistant/components/nice_go/ @IceBotYT
+/tests/components/nice_go/ @IceBotYT
 /homeassistant/components/nightscout/ @marciogranzotto
 /tests/components/nightscout/ @marciogranzotto
 /homeassistant/components/nilu/ @hfurubotten
