+++ homeassistant/components/linear_garage_door/manifest.json
@@ -1,10 +1,9 @@
 {
   "domain": "linear_garage_door",
-  "name": "Nice G.O.",
+  "name": "Linear Garage Door",
   "codeowners": ["@IceBotYT"],
   "config_flow": true,
   "documentation": "https://www.home-assistant.io/integrations/linear_garage_door",
-  "iot_class": "cloud_push",
-  "loggers": ["nice-go"],
-  "requirements": ["nice-go==0.1.6"]
+  "iot_class": "cloud_polling",
+  "requirements": ["linear-garage-door==0.2.9"]
 }
