+++ homeassistant/components/nice_go/cover.py
@@ -1,4 +1,4 @@
-"""Cover entity for Linear Garage Doors."""
+"""Cover entity for Nice G.O."""
 
 from datetime import timedelta
 from typing import Any
@@ -8,37 +8,32 @@ from homeassistant.components.cover import (
     CoverEntity,
     CoverEntityFeature,
 )
-from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from .const import DOMAIN
-from .coordinator import LinearUpdateCoordinator
-from .entity import LinearEntity
+from . import NiceGOConfigEntry
+from .entity import NiceGOEntity
 
-SUPPORTED_SUBDEVICES = ["GDO"]
 PARALLEL_UPDATES = 1
 SCAN_INTERVAL = timedelta(seconds=10)
 
 
 async def async_setup_entry(
     hass: HomeAssistant,
-    config_entry: ConfigEntry,
+    config_entry: NiceGOConfigEntry,
     async_add_entities: AddEntitiesCallback,
 ) -> None:
-    """Set up Linear Garage Door cover."""
-    coordinator: LinearUpdateCoordinator = hass.data[DOMAIN][config_entry.entry_id]
+    """Set up Nice G.O. cover."""
+    coordinator = config_entry.runtime_data
 
     async_add_entities(
-        LinearCoverEntity(coordinator, device_id, device_data.name, sub_device_id)
+        NiceGOCoverEntity(coordinator, device_id, device_data.name, "cover")
         for device_id, device_data in coordinator.data.items()
-        for sub_device_id in device_data.subdevices
-        if sub_device_id in SUPPORTED_SUBDEVICES
     )
 
 
-class LinearCoverEntity(LinearEntity, CoverEntity):
-    """Representation of a Linear cover."""
+class NiceGOCoverEntity(NiceGOEntity, CoverEntity):
+    """Representation of a Nice G.O. cover."""
 
     _attr_supported_features = CoverEntityFeature.OPEN | CoverEntityFeature.CLOSE
     _attr_name = None
@@ -47,41 +42,33 @@ class LinearCoverEntity(LinearEntity, CoverEntity):
     @property
     def is_closed(self) -> bool:
         """Return if cover is closed."""
-        return self.sub_device.get("Open_B") == "false"
+        return self.data.barrier_status == "closed"
 
     @property
     def is_opened(self) -> bool:
         """Return if cover is open."""
-        return self.sub_device.get("Open_B") == "true"
+        return self.data.barrier_status == "open"
 
     @property
     def is_opening(self) -> bool:
         """Return if cover is opening."""
-        return self.sub_device.get("Opening_P") == "0"
+        return self.data.barrier_status == "opening"
 
     @property
     def is_closing(self) -> bool:
         """Return if cover is closing."""
-        return self.sub_device.get("Opening_P") == "100"
+        return self.data.barrier_status == "closing"
 
     async def async_close_cover(self, **kwargs: Any) -> None:
         """Close the garage door."""
         if self.is_closed:
             return
 
-        await self.coordinator.execute(
-            lambda linear: linear.operate_device(
-                self._device_id, self._sub_device_id, "Close"
-            )
-        )
+        await self.coordinator.api.close_barrier(self._device_id)
 
     async def async_open_cover(self, **kwargs: Any) -> None:
         """Open the garage door."""
         if self.is_opened:
             return
 
-        await self.coordinator.execute(
-            lambda linear: linear.operate_device(
-                self._device_id, self._sub_device_id, "Open"
-            )
-        )
+        await self.coordinator.api.open_barrier(self._device_id)
