+++ homeassistant/components/nice_go/diagnostics.py
@@ -1,4 +1,4 @@
-"""Diagnostics support for Linear Garage Door."""
+"""Diagnostics support for Nice G.O.."""
 
 from __future__ import annotations
 
@@ -6,21 +6,20 @@ from dataclasses import asdict
 from typing import Any
 
 from homeassistant.components.diagnostics import async_redact_data
-from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_EMAIL, CONF_PASSWORD
 from homeassistant.core import HomeAssistant
 
-from .const import DOMAIN
-from .coordinator import LinearUpdateCoordinator
+from . import NiceGOConfigEntry
+from .const import CONF_REFRESH_TOKEN
 
-TO_REDACT = {CONF_PASSWORD, CONF_EMAIL}
+TO_REDACT = {CONF_PASSWORD, CONF_EMAIL, CONF_REFRESH_TOKEN, "title"}
 
 
 async def async_get_config_entry_diagnostics(
-    hass: HomeAssistant, entry: ConfigEntry
+    hass: HomeAssistant, entry: NiceGOConfigEntry
 ) -> dict[str, Any]:
     """Return diagnostics for a config entry."""
-    coordinator: LinearUpdateCoordinator = hass.data[DOMAIN][entry.entry_id]
+    coordinator = entry.runtime_data
 
     return {
         "entry": async_redact_data(entry.as_dict(), TO_REDACT),
