+++ homeassistant/components/nice_go/entity.py
@@ -1,20 +1,20 @@
-"""Base entity for Linear."""
+"""Base entity for Nice G.O."""
 
 from homeassistant.helpers.device_registry import DeviceInfo
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import DOMAIN
-from .coordinator import LinearDevice, LinearUpdateCoordinator
+from .coordinator import NiceGODevice, NiceGOUpdateCoordinator
 
 
-class LinearEntity(CoordinatorEntity[LinearUpdateCoordinator]):
-    """Common base for Linear entities."""
+class NiceGOEntity(CoordinatorEntity[NiceGOUpdateCoordinator]):
+    """Common base for Nice G.O. entities."""
 
     _attr_has_entity_name = True
 
     def __init__(
         self,
-        coordinator: LinearUpdateCoordinator,
+        coordinator: NiceGOUpdateCoordinator,
         device_id: str,
         device_name: str,
         sub_device_id: str,
@@ -24,20 +24,13 @@ class LinearEntity(CoordinatorEntity[LinearUpdateCoordinator]):
 
         self._attr_unique_id = f"{device_id}-{sub_device_id}"
         self._device_id = device_id
-        self._sub_device_id = sub_device_id
         self._attr_device_info = DeviceInfo(
             identifiers={(DOMAIN, device_id)},
             name=device_name,
-            manufacturer="Linear",
-            model="Garage Door Opener",
+            sw_version=coordinator.data[device_id].fw_version,
         )
 
     @property
-    def linear_device(self) -> LinearDevice:
-        """Return the Linear device."""
+    def data(self) -> NiceGODevice:
+        """Return the Nice G.O. device."""
         return self.coordinator.data[self._device_id]
-
-    @property
-    def sub_device(self) -> dict[str, str]:
-        """Return the subdevice."""
-        return self.linear_device.subdevices[self._sub_device_id]
