+++ homeassistant/components/nice_go/strings.json
@@ -23,5 +23,11 @@
         "name": "[%key:component::light::title%]"
       }
     }
+  },
+  "issues": {
+    "firmware_update_required": {
+      "title": "Firmware update required",
+      "description": "Your device ({device_name}) requires a firmware update on the Nice G.O. app in order to work with this integration. Please update the firmware on the Nice G.O. app and reconfigure this integration."
+    }
   }
 }
