+++ requirements_all.txt
@@ -1266,6 +1266,9 @@ lightwave==0.24
 # homeassistant.components.limitlessled
 limitlessled==1.1.3
 
+# homeassistant.components.linear_garage_door
+linear-garage-door==0.2.9
+
 # homeassistant.components.linode
 linode-api==4.1.9b1
 
@@ -1416,7 +1419,7 @@ nextdns==3.1.0
 # homeassistant.components.nibe_heatpump
 nibe==2.11.0
 
-# homeassistant.components.linear_garage_door
+# homeassistant.components.nice_go
 nice-go==0.1.6
 
 # homeassistant.components.niko_home_control
