+++ requirements_test_all.txt
@@ -1050,6 +1050,9 @@ librouteros==3.2.0
 # homeassistant.components.soundtouch
 libsoundtouch==0.8
 
+# homeassistant.components.linear_garage_door
+linear-garage-door==0.2.9
+
 # homeassistant.components.lamarzocco
 lmcloud==1.1.13
 
@@ -1176,7 +1179,7 @@ nextdns==3.1.0
 # homeassistant.components.nibe_heatpump
 nibe==2.11.0
 
-# homeassistant.components.linear_garage_door
+# homeassistant.components.nice_go
 nice-go==0.1.6
 
 # homeassistant.components.nfandroidtv
