+++ tests/components/linear_garage_door/fixtures/get_device_state_1.json
@@ -1,0 +1,42 @@
+{
+  "test1": {
+    "GDO": {
+      "Open_B": "true",
+      "Opening_P": "100"
+    },
+    "Light": {
+      "On_B": "false",
+      "On_P": "0"
+    }
+  },
+  "test2": {
+    "GDO": {
+      "Open_B": "false",
+      "Opening_P": "0"
+    },
+    "Light": {
+      "On_B": "true",
+      "On_P": "100"
+    }
+  },
+  "test3": {
+    "GDO": {
+      "Open_B": "false",
+      "Opening_P": "0"
+    },
+    "Light": {
+      "On_B": "false",
+      "On_P": "0"
+    }
+  },
+  "test4": {
+    "GDO": {
+      "Open_B": "true",
+      "Opening_P": "100"
+    },
+    "Light": {
+      "On_B": "true",
+      "On_P": "100"
+    }
+  }
+}
