+++ tests/components/linear_garage_door/fixtures/get_sites.json
@@ -1,0 +1,1 @@
+[{ "id": "test-site-id", "name": "test-site-name" }]
