+++ tests/components/linear_garage_door/snapshots/test_diagnostics.ambr
@@ -2,29 +2,65 @@
 # name: test_entry_diagnostics
   dict({
     'coordinator_data': dict({
-      '1': dict({
-        'barrier_status': 'closed',
-        'connected': True,
-        'fw_version': '1.2.3.4.5.6',
-        'id': '1',
-        'light_status': True,
+      'test1': dict({
         'name': 'Test Garage 1',
+        'subdevices': dict({
+          'GDO': dict({
+            'Open_B': 'true',
+            'Open_P': '100',
+          }),
+          'Light': dict({
+            'On_B': 'true',
+            'On_P': '100',
+          }),
+        }),
       }),
-      '2': dict({
-        'barrier_status': 'open',
-        'connected': True,
-        'fw_version': '1.2.3.4.5.6',
-        'id': '2',
-        'light_status': False,
+      'test2': dict({
         'name': 'Test Garage 2',
+        'subdevices': dict({
+          'GDO': dict({
+            'Open_B': 'false',
+            'Open_P': '0',
+          }),
+          'Light': dict({
+            'On_B': 'false',
+            'On_P': '0',
+          }),
+        }),
+      }),
+      'test3': dict({
+        'name': 'Test Garage 3',
+        'subdevices': dict({
+          'GDO': dict({
+            'Open_B': 'false',
+            'Opening_P': '0',
+          }),
+          'Light': dict({
+            'On_B': 'false',
+            'On_P': '0',
+          }),
+        }),
+      }),
+      'test4': dict({
+        'name': 'Test Garage 4',
+        'subdevices': dict({
+          'GDO': dict({
+            'Open_B': 'true',
+            'Opening_P': '100',
+          }),
+          'Light': dict({
+            'On_B': 'true',
+            'On_P': '100',
+          }),
+        }),
       }),
     }),
     'entry': dict({
       'data': dict({
+        'device_id': 'test-uuid',
         'email': '**REDACTED**',
         'password': '**REDACTED**',
-        'refresh_token': '**REDACTED**',
-        'refresh_token_creation_time': 1722184160.738171,
+        'site_id': 'test-site-id',
       }),
       'disabled_by': None,
       'domain': 'linear_garage_door',
@@ -35,9 +71,9 @@
       'pref_disable_new_entities': False,
       'pref_disable_polling': False,
       'source': 'user',
-      'title': '**REDACTED**',
+      'title': 'test-site-name',
       'unique_id': None,
-      'version': 2,
+      'version': 1,
     }),
   })
 # ---
