+++ tests/components/nice_go/fixtures/device_state_update_1.json
