+++ tests/components/nice_go/fixtures/get_all_barriers.json
