+++ tests/components/nice_go/snapshots/test_diagnostics.ambr
@@ -1,0 +1,43 @@
+# serializer version: 1
+# name: test_entry_diagnostics
+  dict({
+    'coordinator_data': dict({
+      '1': dict({
+        'barrier_status': 'closed',
+        'connected': True,
+        'fw_version': '1.2.3.4.5.6',
+        'id': '1',
+        'light_status': True,
+        'name': 'Test Garage 1',
+      }),
+      '2': dict({
+        'barrier_status': 'open',
+        'connected': True,
+        'fw_version': '1.2.3.4.5.6',
+        'id': '2',
+        'light_status': False,
+        'name': 'Test Garage 2',
+      }),
+    }),
+    'entry': dict({
+      'data': dict({
+        'email': '**REDACTED**',
+        'password': '**REDACTED**',
+        'refresh_token': '**REDACTED**',
+        'refresh_token_creation_time': 1722184160.738171,
+      }),
+      'disabled_by': None,
+      'domain': 'nice_go',
+      'entry_id': 'acefdd4b3a4a0911067d1cf51414201e',
+      'minor_version': 1,
+      'options': dict({
+      }),
+      'pref_disable_new_entities': False,
+      'pref_disable_polling': False,
+      'source': 'user',
+      'title': '**REDACTED**',
+      'unique_id': None,
+      'version': 1,
+    }),
+  })
+# ---
