+++ tests/components/nice_go/snapshots/test_light.ambr
@@ -6,7 +6,7 @@
     'area_id': None,
     'capabilities': dict({
       'supported_color_modes': list([
-        <ColorMode.BRIGHTNESS: 'brightness'>,
+        <ColorMode.ONOFF: 'onoff'>,
       ]),
     }),
     'config_entry_id': <ANY>,
@@ -28,22 +28,21 @@
     'original_device_class': None,
     'original_icon': None,
     'original_name': 'Light',
-    'platform': 'linear_garage_door',
+    'platform': 'nice_go',
     'previous_unique_id': None,
     'supported_features': 0,
     'translation_key': 'light',
-    'unique_id': 'test1-Light',
+    'unique_id': '1-light',
     'unit_of_measurement': None,
   })
 # ---
 # name: test_data[light.test_garage_1_light-state]
   StateSnapshot({
     'attributes': ReadOnlyDict({
-      'brightness': 255,
-      'color_mode': <ColorMode.BRIGHTNESS: 'brightness'>,
+      'color_mode': <ColorMode.ONOFF: 'onoff'>,
       'friendly_name': 'Test Garage 1 Light',
       'supported_color_modes': list([
-        <ColorMode.BRIGHTNESS: 'brightness'>,
+        <ColorMode.ONOFF: 'onoff'>,
       ]),
       'supported_features': <LightEntityFeature: 0>,
     }),
@@ -62,7 +61,7 @@
     'area_id': None,
     'capabilities': dict({
       'supported_color_modes': list([
-        <ColorMode.BRIGHTNESS: 'brightness'>,
+        <ColorMode.ONOFF: 'onoff'>,
       ]),
     }),
     'config_entry_id': <ANY>,
@@ -84,22 +83,21 @@
     'original_device_class': None,
     'original_icon': None,
     'original_name': 'Light',
-    'platform': 'linear_garage_door',
+    'platform': 'nice_go',
     'previous_unique_id': None,
     'supported_features': 0,
     'translation_key': 'light',
-    'unique_id': 'test2-Light',
+    'unique_id': '2-light',
     'unit_of_measurement': None,
   })
 # ---
 # name: test_data[light.test_garage_2_light-state]
   StateSnapshot({
     'attributes': ReadOnlyDict({
-      'brightness': None,
       'color_mode': None,
       'friendly_name': 'Test Garage 2 Light',
       'supported_color_modes': list([
-        <ColorMode.BRIGHTNESS: 'brightness'>,
+        <ColorMode.ONOFF: 'onoff'>,
       ]),
       'supported_features': <LightEntityFeature: 0>,
     }),
