+++ tests/components/nice_go/test_config_flow.py
@@ -1,11 +1,16 @@
-"""Test the Linear Garage Door config flow."""
+"""Test the Nice G.O. config flow."""
 
 from unittest.mock import AsyncMock, patch
 
-from linear_garage_door.errors import InvalidLoginError
+from freezegun.api import FrozenDateTimeFactory
+from nice_go import AuthFailedError
 import pytest
 
-from homeassistant.components.linear_garage_door.const import DOMAIN
+from homeassistant.components.nice_go.const import (
+    CONF_REFRESH_TOKEN,
+    CONF_REFRESH_TOKEN_CREATION_TIME,
+    DOMAIN,
+)
 from homeassistant.config_entries import SOURCE_REAUTH, SOURCE_USER
 from homeassistant.const import CONF_EMAIL, CONF_PASSWORD
 from homeassistant.core import HomeAssistant
@@ -15,7 +20,10 @@ from tests.common import MockConfigEntry
 
 
 async def test_form(
-    hass: HomeAssistant, mock_linear: AsyncMock, mock_setup_entry: AsyncMock
+    hass: HomeAssistant,
+    mock_nice_go: AsyncMock,
+    mock_setup_entry: AsyncMock,
+    freezer: FrozenDateTimeFactory,
 ) -> None:
     """Test we get the form."""
     result = await hass.config_entries.flow.async_init(
@@ -37,27 +45,23 @@ async def test_form(
         )
         await hass.async_block_till_done()
 
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"], {"site": "test-site-id"}
-    )
-    await hass.async_block_till_done()
-
     assert result["type"] is FlowResultType.CREATE_ENTRY
-    assert result["title"] == "test-site-name"
+    assert result["title"] == "test-email"
     assert result["data"] == {
         CONF_EMAIL: "test-email",
         CONF_PASSWORD: "test-password",
-        "site_id": "test-site-id",
-        "device_id": "test-uuid",
+        CONF_REFRESH_TOKEN: "test-refresh-token",
+        CONF_REFRESH_TOKEN_CREATION_TIME: freezer.time_to_freeze.timestamp(),
     }
     assert len(mock_setup_entry.mock_calls) == 1
 
 
 async def test_reauth(
     hass: HomeAssistant,
-    mock_linear: AsyncMock,
+    mock_nice_go: AsyncMock,
     mock_setup_entry: AsyncMock,
     mock_config_entry: MockConfigEntry,
+    freezer: FrozenDateTimeFactory,
 ) -> None:
     """Test reauthentication."""
     mock_config_entry.add_to_hass(hass)
@@ -93,24 +97,24 @@ async def test_reauth(
     assert mock_config_entry.data == {
         CONF_EMAIL: "new-email",
         CONF_PASSWORD: "new-password",
-        "site_id": "test-site-id",
-        "device_id": "test-uuid",
+        CONF_REFRESH_TOKEN: "test-refresh-token",
+        CONF_REFRESH_TOKEN_CREATION_TIME: freezer.time_to_freeze.timestamp(),
     }
 
 
 @pytest.mark.parametrize(
     ("side_effect", "expected_error"),
-    [(InvalidLoginError, "invalid_auth"), (Exception, "unknown")],
+    [(AuthFailedError, "invalid_auth"), (Exception, "unknown")],
 )
 async def test_form_exceptions(
     hass: HomeAssistant,
-    mock_linear: AsyncMock,
+    mock_nice_go: AsyncMock,
     mock_setup_entry: AsyncMock,
     side_effect: Exception,
     expected_error: str,
 ) -> None:
     """Test we handle invalid auth."""
-    mock_linear.login.side_effect = side_effect
+    mock_nice_go.authenticate.side_effect = side_effect
     result = await hass.config_entries.flow.async_init(
         DOMAIN, context={"source": SOURCE_USER}
     )
@@ -125,7 +129,7 @@ async def test_form_exceptions(
 
     assert result["type"] is FlowResultType.FORM
     assert result["errors"] == {"base": expected_error}
-    mock_linear.login.side_effect = None
+    mock_nice_go.authenticate.side_effect = None
     result = await hass.config_entries.flow.async_configure(
         result["flow_id"],
         {
@@ -133,9 +137,5 @@ async def test_form_exceptions(
             CONF_PASSWORD: "test-password",
         },
     )
-    result = await hass.config_entries.flow.async_configure(
-        result["flow_id"], {"site": "test-site-id"}
-    )
-    await hass.async_block_till_done()
 
     assert result["type"] is FlowResultType.CREATE_ENTRY
