+++ tests/components/nice_go/test_cover.py
@@ -1,6 +1,5 @@
-"""Test Linear Garage Door cover."""
+"""Test Nice G.O. cover."""
 
-from datetime import timedelta
 from unittest.mock import AsyncMock
 
 from freezegun.api import FrozenDateTimeFactory
@@ -11,7 +10,7 @@ from homeassistant.components.cover import (
     SERVICE_CLOSE_COVER,
     SERVICE_OPEN_COVER,
 )
-from homeassistant.components.linear_garage_door import DOMAIN
+from homeassistant.components.nice_go.const import DOMAIN
 from homeassistant.const import (
     ATTR_ENTITY_ID,
     STATE_CLOSED,
@@ -25,17 +24,12 @@ from homeassistant.helpers import entity_registry as er
 
 from . import setup_integration
 
-from tests.common import (
-    MockConfigEntry,
-    async_fire_time_changed,
-    load_json_object_fixture,
-    snapshot_platform,
-)
+from tests.common import MockConfigEntry, load_json_object_fixture, snapshot_platform
 
 
 async def test_covers(
     hass: HomeAssistant,
-    mock_linear: AsyncMock,
+    mock_nice_go: AsyncMock,
     entity_registry: er.EntityRegistry,
     snapshot: SnapshotAssertion,
     mock_config_entry: MockConfigEntry,
@@ -48,7 +42,7 @@ async def test_covers(
 
 
 async def test_open_cover(
-    hass: HomeAssistant, mock_linear: AsyncMock, mock_config_entry: MockConfigEntry
+    hass: HomeAssistant, mock_nice_go: AsyncMock, mock_config_entry: MockConfigEntry
 ) -> None:
     """Test that opening the cover works as intended."""
 
@@ -57,24 +51,24 @@ async def test_open_cover(
     await hass.services.async_call(
         COVER_DOMAIN,
         SERVICE_OPEN_COVER,
-        {ATTR_ENTITY_ID: "cover.test_garage_1"},
-        blocking=True,
-    )
-
-    assert mock_linear.operate_device.call_count == 0
-
-    await hass.services.async_call(
-        COVER_DOMAIN,
-        SERVICE_OPEN_COVER,
         {ATTR_ENTITY_ID: "cover.test_garage_2"},
         blocking=True,
     )
 
-    assert mock_linear.operate_device.call_count == 1
+    assert mock_nice_go.open_barrier.call_count == 0
+
+    await hass.services.async_call(
+        COVER_DOMAIN,
+        SERVICE_OPEN_COVER,
+        {ATTR_ENTITY_ID: "cover.test_garage_1"},
+        blocking=True,
+    )
+
+    assert mock_nice_go.open_barrier.call_count == 1
 
 
 async def test_close_cover(
-    hass: HomeAssistant, mock_linear: AsyncMock, mock_config_entry: MockConfigEntry
+    hass: HomeAssistant, mock_nice_go: AsyncMock, mock_config_entry: MockConfigEntry
 ) -> None:
     """Test that closing the cover works as intended."""
 
@@ -83,25 +77,25 @@ async def test_close_cover(
     await hass.services.async_call(
         COVER_DOMAIN,
         SERVICE_CLOSE_COVER,
-        {ATTR_ENTITY_ID: "cover.test_garage_2"},
-        blocking=True,
-    )
-
-    assert mock_linear.operate_device.call_count == 0
-
-    await hass.services.async_call(
-        COVER_DOMAIN,
-        SERVICE_CLOSE_COVER,
         {ATTR_ENTITY_ID: "cover.test_garage_1"},
         blocking=True,
     )
 
-    assert mock_linear.operate_device.call_count == 1
+    assert mock_nice_go.close_barrier.call_count == 0
+
+    await hass.services.async_call(
+        COVER_DOMAIN,
+        SERVICE_CLOSE_COVER,
+        {ATTR_ENTITY_ID: "cover.test_garage_2"},
+        blocking=True,
+    )
+
+    assert mock_nice_go.close_barrier.call_count == 1
 
 
 async def test_update_cover_state(
     hass: HomeAssistant,
-    mock_linear: AsyncMock,
+    mock_nice_go: AsyncMock,
     mock_config_entry: MockConfigEntry,
     freezer: FrozenDateTimeFactory,
 ) -> None:
@@ -109,16 +103,13 @@ async def test_update_cover_state(
 
     await setup_integration(hass, mock_config_entry, [Platform.COVER])
 
-    assert hass.states.get("cover.test_garage_1").state == STATE_OPEN
-    assert hass.states.get("cover.test_garage_2").state == STATE_CLOSED
+    assert hass.states.get("cover.test_garage_1").state == STATE_CLOSED
+    assert hass.states.get("cover.test_garage_2").state == STATE_OPEN
 
-    device_states = load_json_object_fixture("get_device_state_1.json", DOMAIN)
-    mock_linear.get_device_state.side_effect = lambda device_id: device_states[
-        device_id
-    ]
+    device_update = load_json_object_fixture("device_state_update.json", DOMAIN)
+    await mock_config_entry.runtime_data.on_data(device_update)
+    device_update_1 = load_json_object_fixture("device_state_update_1.json", DOMAIN)
+    await mock_config_entry.runtime_data.on_data(device_update_1)
 
-    freezer.tick(timedelta(seconds=60))
-    async_fire_time_changed(hass)
-
-    assert hass.states.get("cover.test_garage_1").state == STATE_CLOSING
-    assert hass.states.get("cover.test_garage_2").state == STATE_OPENING
+    assert hass.states.get("cover.test_garage_1").state == STATE_OPENING
+    assert hass.states.get("cover.test_garage_2").state == STATE_CLOSING
